Refactor and update dependencies
- 🗑️ Remove .npmrc file due to pnpm issues
- ✨ Add pre-commit hook for type checking
- 🔧 Update pnpm workspace configuration
- 📦 Update dependencies:
  - nuxt: 3.15.4 -> 3.16.0
  - pnpm: 10.5.2 -> 10.6.1
  - @cloudflare/vitest-pool-workers: 0.7.6 -> 0.7.7
  - vitest: 3.0.7 -> 3.0.8
- 🔄 Refactor Twitch authentication utility functions
