+++ .husky/pre-commit
@@ -1,0 +1,1 @@
+pnpm --filter @pepega/twitch-webhooks --filter @pepega/pepega run typecheck
