--- .npmrc
@@ -1,2 +1,0 @@
-# Something wrong with pnpm and Nuxt build without this
-node-linker=hoisted
