+++ apps/pepega/package.json
@@ -25,7 +25,7 @@
     "date-fns": "4.1.0",
     "drizzle-orm": "catalog:",
     "modern-normalize": "3.0.1",
-    "nuxt": "3.15.4",
+    "nuxt": "3.16.0",
     "pinia": "3.0.1",
     "tsx": "4.19.3",
     "ufo": "1.5.4",
