+++ common/twitch/package.json
@@ -15,6 +15,7 @@
   },
   "dependencies": {
     "@pepega/utils": "workspace:*",
-    "ofetch": "1.4.1"
+    "ofetch": "1.4.1",
+    "valibot": "1.0.0-rc.3"
   }
 }
