+++ pnpm-workspace.yaml
@@ -15,3 +15,7 @@ onlyBuiltDependencies:
   - esbuild
   - sharp
   - workerd
+
+# .npmrc
+# Something wrong with pnpm and Nuxt build without this
+nodeLinker: hoisted
