+++ workers/twitch-webhooks/package.json
@@ -8,10 +8,11 @@
     "dev": "wrangler dev",
     "start": "wrangler dev",
     "test": "vitest",
-    "cf-typegen": "wrangler types"
+    "cf-typegen": "wrangler types",
+    "typecheck": "tsc --noEmit"
   },
   "dependencies": {
-    "@cloudflare/vitest-pool-workers": "0.7.6",
+    "@cloudflare/vitest-pool-workers": "0.7.7",
     "@cloudflare/workers-types": "4.20250303.0",
     "@pepega/database": "workspace:*",
     "@pepega/twitch": "workspace:*",
@@ -21,7 +22,7 @@
     "h3": "1.15.1",
     "typescript": "catalog:",
     "valibot": "1.0.0-rc.3",
-    "vitest": "3.0.7",
+    "vitest": "3.0.8",
     "wrangler": "catalog:"
   }
 }
